Added code to control two DC motors with TB6612 motor drivers
This code includes both MicroPython code and Fritzing drawings. The
drawings include both the Adafruit and Sparkfun versions of the TB6612
breakout boards. There are both "direct" control of the pyboard pins in
the main script and use of `TB6612.py` which contains a convenience
class for the driver that abstracts away most of the low-level control.